Moved tracing code to separate module
